# Faster symmetric_interval_hull for ExponentialMap of singletons and hyperrectangles