--- a/src/Approximations/Approximations.jl
+++ b/src/Approximations/Approximations.jl
@@ -10,7 +10,7 @@ using LazySets, LazySets.Arrays, Requires, LinearAlgebra, SparseArrays,
       MathProgBase
 
 using LazySets: _isapprox, _leq, _geq, _rtol, _normal_Vector, isapproxzero,
-                default_lp_solver, _isbounded_stiemke
+                default_lp_solver, _isbounded_stiemke, require
 
 import LazySets: project