--- a/src/Approximations/init_Expokit.jl
+++ b/src/Approximations/init_Expokit.jl
@@ -1,0 +1,9 @@
+function load_expokit()
+return quote
+
+using .Expokit: expmv
+
+end end  # quote / load_expokit
+
+
+eval(load_expokit())