--- a/src/Approximations/symmetric_interval_hull.jl
+++ b/src/Approximations/symmetric_interval_hull.jl
@@ -98,3 +98,27 @@ function symmetric_interval_hull(lm::LinearMap{N, <:AbstractHyperrectangle}) whe
     r = abs.(M * center(H)) .+ abs.(M) * radius_hyperrectangle(H)
     return Hyperrectangle(zeros(N, n), r)
 end
+
+function symmetric_interval_hull(E::ExponentialMap{N, <:AbstractSingleton}) where {N}
+    require(:Expokit; fun_name="symmetric_interval_hull")
+
+    v = expmv(one(N), E.spmexp.M, element(E.X))
+    c = zeros(N, dim(E))
+    r = abs.(v)
+    return Hyperrectangle(c, r)
+end
+
+function symmetric_interval_hull(E::ExponentialMap{N, <:AbstractHyperrectangle}) where {N}
+    require(:Expokit; fun_name="symmetric_interval_hull")
+
+    H = set(E)
+    n = dim(H)
+    x = zeros(N, n)
+    r = abs.(expmv(one(N), E.spmexp.M, center(H)))
+    @inbounds for i in 1:n
+        x[i] = radius_hyperrectangle(H, i)
+        r .+= abs.(expmv(one(N), E.spmexp.M, x))
+        x[i] = 0
+    end
+    return Hyperrectangle(zeros(N, n), r)
+end