--- a/test/Approximations/symmetric_interval_hull.jl
+++ b/test/Approximations/symmetric_interval_hull.jl
@@ -18,3 +18,23 @@ for N in [Float64, Rational{Int}, Float32]
     H2 = Hyperrectangle(zeros(N, 4), N[6, 7, 16, 18])
     @test symmetric_interval_hull(M * H1) == H2
 end
+
+# tests that only work with Float64 and Float32
+for N in [Float64, Float32]
+    # exponential map of singleton
+    M = sparse(diagm(N[1, 1, 1]))
+    E = SparseMatrixExp(M) * Singleton(N[1, 2, 3])
+    H = Hyperrectangle(zeros(N, 3), N[ℯ, 2ℯ, 3ℯ])
+    @test symmetric_interval_hull(E) ≈ H
+
+    # exponential map of hyperrectangle
+    M = sparse(diagm(N[1, 1, 1]))
+    E = SparseMatrixExp(M) * Hyperrectangle(N[1, 2, 3], N[1, 2, 3])
+    H = symmetric_interval_hull(E)
+    @test H ≈ Hyperrectangle(zeros(N, 3), N[2ℯ, 4ℯ, 6ℯ])
+    # matrix with negative entries
+    M = sparse(diagm(N[1, -1]))
+    E = SparseMatrixExp(M) * Hyperrectangle(N[1, 1], N[1, 1])
+    H = symmetric_interval_hull(E)
+    @test H ≈ Hyperrectangle(zeros(N, 2), N[2ℯ, 2ℯ^(-1)])
+end