Using `Expokit` in `Approximations` required a small refactoring.

```julia
using LazySets, SparseArrays, Expokit, Test

for n in [1, 10, 100, 200]
    A = sparse(rand(n, n) .- rand(n, n));
    δ = 0.1;
    M = SparseMatrixExp(A * δ);

    println("n = $n\nSingleton")

    X = rand(Singleton, dim=n);
    EX = ExponentialMap(M, X);
    @time SX = symmetric_interval_hull(EX);  # master
    @time SX2 = LazySets.Approximations.symmetric_interval_hull2(EX);  # this branch
    @test isequivalent(SX, SX2)

    println("Hyperrectangle")

    Y = rand(Hyperrectangle, dim=n);
    EY = ExponentialMap(M, Y);
    @time SY = symmetric_interval_hull(EY);  # master
    @time SY2 = LazySets.Approximations._symmetric_interval_hull_krylov(EY);  # this branch
    @time SY3 = LazySets.Approximations._symmetric_interval_hull_exp(EY);  # this branch
    @test isequivalent(SY, SY2)
    @test isequivalent(SY, SY3)
end
```
---
```julia
n = 1
Singleton
  0.000094 seconds (119 allocations: 15.000 KiB)
  0.000028 seconds (57 allocations: 7.344 KiB)
Hyperrectangle
  0.000047 seconds (119 allocations: 15.000 KiB)
  0.000043 seconds (113 allocations: 14.594 KiB)
  0.000020 seconds (26 allocations: 2.281 KiB)
n = 10
Singleton
  0.002171 seconds (1.54 k allocations: 1.191 MiB)
  0.000106 seconds (78 allocations: 61.234 KiB)
Hyperrectangle
  0.002242 seconds (1.54 k allocations: 1.191 MiB)
  0.001896 seconds (839 allocations: 670.609 KiB)
  0.000102 seconds (37 allocations: 23.969 KiB)
n = 100
Singleton
  0.284080 seconds (14.01 k allocations: 53.663 MiB, 5.15% gc time)
  0.001111 seconds (71 allocations: 276.438 KiB)
Hyperrectangle
  0.300344 seconds (14.01 k allocations: 53.663 MiB, 2.81% gc time)
  0.120292 seconds (6.97 k allocations: 27.096 MiB)
  0.001231 seconds (43 allocations: 1.227 MiB)
n = 200
Singleton
  3.501998 seconds (32.97 k allocations: 149.301 MiB, 0.58% gc time)
  0.006972 seconds (83 allocations: 382.453 KiB)
Hyperrectangle
  4.091444 seconds (32.97 k allocations: 149.301 MiB, 0.45% gc time)
  1.412295 seconds (15.62 k allocations: 69.847 MiB, 0.50% gc time)
  0.006196 seconds (47 allocations: 5.505 MiB)
```

The first output is `master`. The second output is the Krylov approach. For hyperrectangles the direct computation of the matrix exponential (third output) is much faster (as long that is possible), so I used this one by default.