@mforets 2021-08-12
yes, i agree to only use `_symmetric_interval_hull_krylov(E::ExponentialMap{N, <:AbstractHyperrectangle})` in this case